requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
pytest==8.2.1
behave==1.2.6
loguru==0.7.2
allure-behave==2.13.5
pillow==10.3.0
requests==2.32.2
behave-html-formatter==0.9.10
allure-combine==1.0.11